Create separate file for `TermTile`
